waf: fix profile caching, score overflow, and remove expect panic
Bug Fixes:
- profiles.rs: get_waf_profiles() now uses LazyLock static cache
  instead of recreating Vec on every call; get_profile_by_name()
  returns cloned profile from cached reference
- mod.rs: replaced expect() panic with proper if let + early return
- constants.rs: changed scoring constants from u8 to u16 to prevent
  overflow when accumulating multiple matches
- detect.rs: changed score accumulator and best_match to u16 to
  match constant types; added cast to u8 when storing confidence

Documentation:
- Updated AGENTS.md with new WAF bug fixes section
- Updated architecture/waf.md with scoring overflow note and profile
  caching info
- Updated AGENTS.override.md with Profile Caching and Score Overflow
  Prevention sections
- Updated slapper-waf SKILL.md with u16 scoring note and profile
  caching guidance

+++ crates/slapper/src/constants.rs
@@ -67,13 +67,13 @@ pub mod ui {
 }
 
 pub mod waf {
-    pub const HEADER_MATCH_SCORE: u8 = 25;
-    pub const COOKIE_MATCH_SCORE: u8 = 20;
-    pub const BODY_MATCH_SCORE: u8 = 15;
-    pub const IP_MATCH_SCORE: u8 = 20;
-    pub const UNKNOWN_WAF_CONFIDENCE: u8 = 30;
+    pub const HEADER_MATCH_SCORE: u16 = 25;
+    pub const COOKIE_MATCH_SCORE: u16 = 20;
+    pub const BODY_MATCH_SCORE: u16 = 15;
+    pub const IP_MATCH_SCORE: u16 = 20;
+    pub const UNKNOWN_WAF_CONFIDENCE: u16 = 30;
     pub const LENGTH_DIFF_THRESHOLD: usize = 100;
-    pub const HIGH_CONFIDENCE_EXIT: u8 = 90;
+    pub const HIGH_CONFIDENCE_EXIT: u16 = 90;
     pub const BLOCKED_STATUS_CODES: [u16; 4] = [403, 406, 429, 503];
     pub const BLOCKED_PATTERNS: [&str; 8] = [
         "access denied",
